Move the snowbridge-runtime-common
